Reordered chat and message table
Change is intended to get around behaviour of the code when mapping and offsetting referenced ids.
Intended to go with Main.kt Circular "capable" version 2.2

Because Mapping is done before processing the tables which identifies uniques, the values are offset too far because those entries duplicated unique indexes are still there and hence counted for offset. By processing `chat` first, the chance of a potentially in accurate mapping is lower as `message` should not contain entries which have the unique indexes clashing. The only scenario this will happen is if there are duplicated msgs. Whereas with `chat` it is gauranteed to contain a clash if you have continuously interacted with at least 1 contact and had new ones as well.

Effect: In the previous order chat_row_ids will be in accurate if there were lines are skipped due to unique indexes set for `chat` table, where the likelihood is much higher than `message` table.