Offset increased by 1
Increased the amount of offset by1 to avoid overlap due to `message` table behaviour listing 0 as jid when it is one-to-one chats and for messages from self. Where before all those jid will overlap with the last jid in the first database. As of now it doesnt affect the display of those messages as all the other infromation makes it display correctly, but would it break in the future? who knows.

This change makes it easier for the user to search for those jids that should be 0 be searching for the id that will be reported as missing in the output consistency check. They can then fix it if they like.

The user needs to change the `sort_id` for the first line in `message` table to 1 from 0, otherwise you will not get the full output consistency check as it will fail on this column.